Split timeline component
