+++ src/components/TimeLine/index.tsx
@@ -1,0 +1,2 @@
+export { Note } from "./Note";
+export { TimeLine } from "./TimeLine";
